Move test files and add new test launch config
