+++ src/test/integration/gradle-multi-project/formatting.test.ts
@@ -13,9 +13,9 @@ import {
   multiProjectJavaFormattedAppFileContents,
   multiProjectJavaHelloFilePath,
   multiProjectJavaHelloFileContents,
-} from '../../../testUtil';
-import { SPOTLESS_GRADLE_EXTENSION_ID } from '../../../../constants';
-import { ExtensionApi } from '../../../../extension';
+} from '../../testUtil';
+import { SPOTLESS_GRADLE_EXTENSION_ID } from '../../../constants';
+import { ExtensionApi } from '../../../extension';
 
 describe('Formatting', () => {
   const { logger, spotless } = vscode.extensions.getExtension(
