+++ src/test/integration/gradle-multi-project/index.ts
