+++ src/test/integration/gradle-project/index.ts
